Merge from upstream

+++ test/test_dataloader.py
@@ -205,9 +205,12 @@ class SleepDataset(Dataset):
     def __init__(self, size, sleep_sec):
         self.size = size
         self.sleep_sec = sleep_sec
+        self.sleeped = False
 
     def __getitem__(self, idx):
-        time.sleep(self.sleep_sec)
+        if not self.sleeped:
+            time.sleep(self.sleep_sec)
+            self.sleeped = True
         return idx
 
     def __len__(self):
@@ -251,7 +254,7 @@ class SynchronizedSeedDataset(Dataset):
 
 
 def _test_timeout():
-    dataset = SleepDataset(10, 10)
+    dataset = SleepDataset(10, 3)
     dataloader = DataLoader(dataset, batch_size=2, num_workers=2, timeout=1)
     _ = next(iter(dataloader))
 
@@ -478,36 +481,40 @@ class TestDataLoader(TestCase):
     @unittest.skipIf(not TEST_CUDA, "CUDA unavailable")
     @skipIfRocm
     def test_partial_workers(self):
-        "check that workers exit even if the iterator is not exhausted"
-        loader = iter(DataLoader(self.dataset, batch_size=2, num_workers=4, pin_memory=True))
-        workers = loader.workers
-        worker_manager_thread = loader.worker_manager_thread
-        for i, sample in enumerate(loader):
-            if i == 3:
-                break
-        del loader
-        for w in workers:
-            w.join(JOIN_TIMEOUT)
-            self.assertFalse(w.is_alive(), 'subprocess not terminated')
-            self.assertEqual(w.exitcode, 0)
-        worker_manager_thread.join(JOIN_TIMEOUT)
-        self.assertFalse(worker_manager_thread.is_alive())
+        r"""Check that workers exit even if the iterator is not exhausted."""
+        for pin_memory in (True, False):
+            loader = iter(DataLoader(self.dataset, batch_size=2, num_workers=4, pin_memory=pin_memory))
+            workers = loader.workers
+            if pin_memory:
+                pin_memory_thread = loader.pin_memory_thread
+            for i, sample in enumerate(loader):
+                if i == 10:
+                    break
+            del loader
+            for w in workers:
+                w.join(JOIN_TIMEOUT)
+                self.assertFalse(w.is_alive(), 'subprocess not terminated')
+            if pin_memory:
+                pin_memory_thread.join(JOIN_TIMEOUT)
+                self.assertFalse(pin_memory_thread.is_alive())
 
     @staticmethod
-    def _manager_process(dataset, worker_pids, manager_exit_event):
+    def _main_process(dataset, worker_pids, main_exit_event, raise_error):
         loader = iter(DataLoader(dataset, batch_size=2, num_workers=4, pin_memory=True))
         workers = loader.workers
         for i in range(len(workers)):
             worker_pids[i] = int(workers[i].pid)
         for i, sample in enumerate(loader):
             if i == 3:
-                break
-        # Simulate a dirty exit of the manager process
-        manager_exit_event.set()
-        if IS_WINDOWS:
-            os.system('taskkill /PID ' + str(os.getpid()) + ' /F')
-        else:
-            os.kill(os.getpid(), signal.SIGKILL)
+                # Simulate an exit of the manager process
+                main_exit_event.set()
+                if raise_error:
+                    raise RuntimeError('Error')
+                else:
+                    if IS_WINDOWS:
+                        os.system('taskkill /PID ' + str(os.getpid()) + ' /F')
+                    else:
+                        os.kill(os.getpid(), signal.SIGKILL)
 
     @staticmethod
     def _is_process_alive(pid, pname):
@@ -532,32 +539,43 @@ class TestDataLoader(TestCase):
     @unittest.skipIf(not TEST_CUDA, "CUDA unavailable")
     @skipIfRocm
     def test_main_process_unclean_exit(self):
-        '''There might be ConnectionResetError or leaked semaphore warning (due to dirty process exit), \
+        r'''There might be ConnectionResetError or leaked semaphore warning (due to dirty process exit), \
 but they are all safe to ignore'''
-        worker_pids = mp.Array('i', [0] * 4)
 
-        manager_exit_event = mp.Event()
-        p = mp.Process(target=TestDataLoader._manager_process,
-                       args=(self.dataset, worker_pids, manager_exit_event))
-        p.start()
+        # `raise_error` controls if the main process is KILL-ed by OS or just
+        # simply raises an error. Both cases are interesting because
+        # 1. In case of it is KILL-ed by OS, the workers need to automatically
+        #    discover that their parent is dead and exit gracefully.
+        # 2. In case of it raises an error itself, the parent process needs to
+        #    take care of exiting the worker and then exits itself gracefully.
+        for raise_error in (True, False):
+            worker_pids = mp.Array('i', [0] * 4)
 
-        manager_exit_event.wait()
+            main_exit_event = mp.Event()
+            p = mp.Process(target=TestDataLoader._main_process,
+                           args=(self.dataset, worker_pids, main_exit_event, raise_error))
+            p.start()
+            worker_pids[-1] = p.pid
 
-        exit_status = [False] * len(worker_pids)
-        start_time = time.time()
-        pname = 'python'
-        while True:
-            for i in range(len(worker_pids)):
-                pid = worker_pids[i]
-                if not exit_status[i]:
-                    if not TestDataLoader._is_process_alive(pid, pname):
-                        exit_status[i] = True
-            if all(exit_status):
-                break
-            else:
-                time.sleep(1)
-                self.assertFalse(time.time() - start_time > MANAGER_STATUS_CHECK_INTERVAL + JOIN_TIMEOUT,
-                                 'subprocess not terminated')
+            main_exit_event.wait()
+
+            exit_status = [False] * len(worker_pids)
+            start_time = time.time()
+            pname = 'python'
+            while True:
+                for i in range(len(worker_pids)):
+                    pid = worker_pids[i]
+                    if not exit_status[i]:
+                        if not TestDataLoader._is_process_alive(pid, pname):
+                            exit_status[i] = True
+                if all(exit_status):
+                    break
+                else:
+                    if time.time() - start_time > MANAGER_STATUS_CHECK_INTERVAL + JOIN_TIMEOUT:
+                        self.fail('subprocess not terminated')
+                    time.sleep(1)
+            p.join(MANAGER_STATUS_CHECK_INTERVAL + JOIN_TIMEOUT - (time.time() - start_time))
+            self.assertFalse(p.is_alive(), 'main process not terminated')
 
     def test_len(self):
         def check_len(dl, expected):
@@ -601,7 +619,7 @@ but they are all safe to ignore'''
             self.assertIsInstance(batch, tt)
 
     @unittest.skipIf(not TEST_NUMPY, "numpy unavailable")
-    def test_default_colate_bad_numpy_types(self):
+    def test_default_collate_bad_numpy_types(self):
         import numpy as np
 
         # Should be a no-op
